App Web DataEmotion/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
fsspec==2024.3.1
huggingface-hub==0.23.0
idna==3.7
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
six==1.16.0
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.40.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
